model/requirements.txt
torch
transformers
pillow
torchvision
pytest
matplotlibpandas
tqdm
